backend/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
cryptography==43.0.1
debugpy==1.8.6
decorator==5.1.1
defusedxml==0.8.0rc2
Django==5.1.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-oauth-toolkit==3.0.1
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-social-oauth2==3.1.0
drf-yasg==1.21.7
exceptiongroup==1.2.2
executing==2.1.0
idna==3.10
inflection==0.5.1
ipython==8.27.0
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jwcrypto==1.5.6
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
nltk==3.9.1
oauthlib==3.2.2
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
stack-data==0.6.3
textblob==0.18.0.post0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
